Cap SDL joystick count to `MAX_JOYSTICKS`
`JoyList` is allocated based on `MAX_JOYSTICKS`, but it's then populated the number of times returned by `SDL_NumJoysticks` without capping it to the constant. This could cause an issue if the user has more than... 8 controllers connected. Not likely, especially since this unit doesn't support controller selection, but it does mean I can eliminate a check in `scanJoysticks` that was too late and is now redundant.